+++ pytest_bdd/plugin.py
@@ -2,11 +2,11 @@
 
 import pytest
 
-from . import given, when, then
 from . import cucumber_json
 from . import generation
-from . import reporting
 from . import gherkin_terminal_reporter
+from . import given, when, then
+from . import reporting
 from .utils import CONFIG_STACK
 
 
@@ -84,25 +84,3 @@ def pytest_cmdline_main(config):
 def pytest_bdd_apply_tag(tag, function):
     mark = getattr(pytest.mark, tag)
     return mark(function)
-
-
-@pytest.mark.tryfirst
-def pytest_collection_modifyitems(session, config, items):
-    """Re-order items using the creation counter as fallback.
-
-    Pytest has troubles to correctly order the test items for python < 3.6.
-    For this reason, we have to apply some better ordering for pytest_bdd scenario-decorated test functions.
-
-    This is not needed for python 3.6+, but this logic is safe to apply in that case as well.
-    """
-    # TODO: Try to only re-sort the items that have __pytest_bdd_counter__, and not the others,
-    #  since there may be other hooks that are executed before this and that want to reorder item as well
-    def item_key(item):
-        if isinstance(item, pytest.Function):
-            declaration_order = getattr(item.function, "__pytest_bdd_counter__", 0)
-        else:
-            declaration_order = 0
-        func, linenum = item.reportinfo()[:2]
-        return (func, linenum if linenum is not None else -1, declaration_order)
-
-    items.sort(key=item_key)