+++ pytest_bdd/scripts.py
@@ -5,7 +5,6 @@ import os.path
 import re
 
 import glob2
-import six
 
 from .generation import generate_code, parse_feature_files
 
@@ -49,10 +48,7 @@ def print_generated_code(args):
     """Print generated test code for the given filenames."""
     features, scenarios, steps = parse_feature_files(args.files)
     code = generate_code(features, scenarios, steps)
-    if six.PY2:
-        print(code.encode("utf-8"))
-    else:
-        print(code)
+    print(code)
 
 
 def main():