+++ tests/feature/test_report.py
@@ -1,5 +1,6 @@
 """Test scenario reporting."""
 import textwrap
+
 import pytest
 
 
@@ -103,22 +104,22 @@ def test_step_trace(testdir):
     report = result.matchreport("test_passing", when="call").scenario
     expected = {
         "feature": {
-            "description": u"",
+            "description": "",
             "filename": feature.strpath,
             "line_number": 2,
-            "name": u"One passing scenario, one failing scenario",
+            "name": "One passing scenario, one failing scenario",
             "rel_filename": relpath,
-            "tags": [u"feature-tag"],
+            "tags": ["feature-tag"],
         },
         "line_number": 5,
-        "name": u"Passing",
+        "name": "Passing",
         "steps": [
             {
                 "duration": OfType(float),
                 "failed": False,
                 "keyword": "Given",
                 "line_number": 6,
-                "name": u"a passing step",
+                "name": "a passing step",
                 "type": "given",
             },
             {
@@ -126,11 +127,11 @@ def test_step_trace(testdir):
                 "failed": False,
                 "keyword": "And",
                 "line_number": 7,
-                "name": u"some other passing step",
+                "name": "some other passing step",
                 "type": "given",
             },
         ],
-        "tags": [u"scenario-passing-tag"],
+        "tags": ["scenario-passing-tag"],
         "examples": [],
         "example_kwargs": {},
     }
@@ -140,22 +141,22 @@ def test_step_trace(testdir):
     report = result.matchreport("test_failing", when="call").scenario
     expected = {
         "feature": {
-            "description": u"",
+            "description": "",
             "filename": feature.strpath,
             "line_number": 2,
-            "name": u"One passing scenario, one failing scenario",
+            "name": "One passing scenario, one failing scenario",
             "rel_filename": relpath,
-            "tags": [u"feature-tag"],
+            "tags": ["feature-tag"],
         },
         "line_number": 10,
-        "name": u"Failing",
+        "name": "Failing",
         "steps": [
             {
                 "duration": OfType(float),
                 "failed": False,
                 "keyword": "Given",
                 "line_number": 11,
-                "name": u"a passing step",
+                "name": "a passing step",
                 "type": "given",
             },
             {
@@ -163,11 +164,11 @@ def test_step_trace(testdir):
                 "failed": True,
                 "keyword": "And",
                 "line_number": 12,
-                "name": u"a failing step",
+                "name": "a failing step",
                 "type": "given",
             },
         ],
-        "tags": [u"scenario-failing-tag"],
+        "tags": ["scenario-failing-tag"],
         "examples": [],
         "example_kwargs": {},
     }
@@ -176,22 +177,22 @@ def test_step_trace(testdir):
     report = result.matchreport("test_outlined[12-5.0-7]", when="call").scenario
     expected = {
         "feature": {
-            "description": u"",
+            "description": "",
             "filename": feature.strpath,
             "line_number": 2,
-            "name": u"One passing scenario, one failing scenario",
+            "name": "One passing scenario, one failing scenario",
             "rel_filename": relpath,
-            "tags": [u"feature-tag"],
+            "tags": ["feature-tag"],
         },
         "line_number": 14,
-        "name": u"Outlined",
+        "name": "Outlined",
         "steps": [
             {
                 "duration": OfType(float),
                 "failed": False,
                 "keyword": "Given",
                 "line_number": 15,
-                "name": u"there are <start> cucumbers",
+                "name": "there are <start> cucumbers",
                 "type": "given",
             },
             {
@@ -199,7 +200,7 @@ def test_step_trace(testdir):
                 "failed": False,
                 "keyword": "When",
                 "line_number": 16,
-                "name": u"I eat <eat> cucumbers",
+                "name": "I eat <eat> cucumbers",
                 "type": "when",
             },
             {
@@ -207,7 +208,7 @@ def test_step_trace(testdir):
                 "failed": False,
                 "keyword": "Then",
                 "line_number": 17,
-                "name": u"I should have <left> cucumbers",
+                "name": "I should have <left> cucumbers",
                 "type": "then",
             },
         ],
@@ -227,22 +228,22 @@ def test_step_trace(testdir):
     report = result.matchreport("test_outlined[5-4.0-1]", when="call").scenario
     expected = {
         "feature": {
-            "description": u"",
+            "description": "",
             "filename": feature.strpath,
             "line_number": 2,
-            "name": u"One passing scenario, one failing scenario",
+            "name": "One passing scenario, one failing scenario",
             "rel_filename": relpath,
-            "tags": [u"feature-tag"],
+            "tags": ["feature-tag"],
         },
         "line_number": 14,
-        "name": u"Outlined",
+        "name": "Outlined",
         "steps": [
             {
                 "duration": OfType(float),
                 "failed": False,
                 "keyword": "Given",
                 "line_number": 15,
-                "name": u"there are <start> cucumbers",
+                "name": "there are <start> cucumbers",
                 "type": "given",
             },
             {
@@ -250,7 +251,7 @@ def test_step_trace(testdir):
                 "failed": False,
                 "keyword": "When",
                 "line_number": 16,
-                "name": u"I eat <eat> cucumbers",
+                "name": "I eat <eat> cucumbers",
                 "type": "when",
             },
             {
@@ -258,7 +259,7 @@ def test_step_trace(testdir):
                 "failed": False,
                 "keyword": "Then",
                 "line_number": 17,
-                "name": u"I should have <left> cucumbers",
+                "name": "I should have <left> cucumbers",
                 "type": "then",
             },
         ],